refactor: `selectedItems` jotai 전역 변수로 대체
