+++ app/constants/index.ts
@@ -41,5 +41,3 @@ export const DISALLOWED_USERNAME_LIST = [
   'wanker',
   'whore',
 ];
-
-export const SELECTED_ITEMS_KEY = 'items';
