+++ app/ui/bookshelf/book-thumbnail.tsx
@@ -2,19 +2,38 @@
 
 import Link from 'next/link';
 import Image from 'next/image';
-import { useAtomValue } from 'jotai';
+import { useState, useEffect } from 'react';
+import { useAtomValue, useSetAtom } from 'jotai';
 import { CheckCircleIcon } from '@heroicons/react/24/outline';
 import InvalidThumbnail from '@/ui//invalid-thumbnail';
-import { useBookThumbnail } from '@/utils/hooks/useBookThumbnail';
 import { getImageSize } from '@/utils/image';
-import { currentModeAtom } from '@/store/atoms';
+import { currentModeAtom, selectedItemsAtom } from '@/store/atoms';
 
 type Props = { id: number; thumbnail: string; title: string };
 
 export default function BookThumbnail({ id, thumbnail, title }: Props) {
   const { width, height } = getImageSize(thumbnail);
-  const { isSelected, handleClick } = useBookThumbnail(id, thumbnail);
   const currentMode = useAtomValue(currentModeAtom);
+  const [isSelected, setIsSelected] = useState(false);
+  const setSelectedItems = useSetAtom(selectedItemsAtom);
+
+  useEffect(() => {
+    setIsSelected(false);
+  }, [currentMode]);
+
+  const handleClick = () => {
+    // 로컬 상태 isSelected 값 업데이트
+    setIsSelected((prev) => !prev);
+
+    // 전역 상태 selectedItems 값 업데이트
+    if (isSelected) {
+      // id 삭제
+      setSelectedItems((prev) => prev.filter((itemId) => itemId !== id));
+    } else {
+      // id 추가
+      setSelectedItems((prev) => [...prev, id]);
+    }
+  };
 
   if (!width || !height) {
     return <InvalidThumbnail />;
